+++ ephios/plugins/basesignup/signals.py
@@ -1,8 +1,10 @@
 from django.dispatch import receiver
 
 from ephios.event_management.signup import register_signup_methods
-from ephios.plugins.basesignup.signup.confirm import RequestConfirmSignupMethod
-from ephios.plugins.basesignup.signup.instant import InstantConfirmationSignupMethod
+from ephios.plugins.basesignup.signup import (
+    InstantConfirmationSignupMethod,
+    RequestConfirmSignupMethod,
+)
 
 
 @receiver(
