+++ ephios/plugins/basesignup/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from ephios.plugins.basesignup.signup.confirm import RequestConfirmDispositionView
+from ephios.plugins.basesignup.signup import RequestConfirmDispositionView
 
 app_name = "basesignup"
 urlpatterns = [
